requirements.txt
fastapi
uvicorn
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
streamlit
requests
langchain
langchain-groq
